ns: Move `NamespaceResolver` from module `reader` to module `name`
This is quite big part of library that needed its own tests,
so it is better to it to live in a dedicated module
