+++ src/name.rs
@@ -1,0 +1,208 @@
+//! Module for handling names according to the W3C [Namespaces in XML 1.1 (Second Edition)][spec]
+//! specification
+//!
+//! [spec]: https://www.w3.org/TR/xml-names11
+
+use crate::events::attributes::Attribute;
+use crate::events::BytesStart;
+
+/// An entry that contains index into the buffer with namespace bindings.
+///
+/// Defines a mapping from *[namespace prefix]* to *[namespace name]*.
+/// If prefix is empty, defines a *default namespace* binding that applies to
+/// unprefixed element names (unprefixed attribute names do not bind to any
+/// namespace and they processing is dependent on the element in which their
+/// defined).
+///
+/// [namespace prefix]: https://www.w3.org/TR/xml-names11/#dt-prefix
+/// [namespace name]: https://www.w3.org/TR/xml-names11/#dt-NSName
+#[derive(Debug, Clone)]
+struct NamespaceEntry {
+    /// Index of the namespace in the buffer
+    start: usize,
+    /// Length of the prefix
+    /// * if greater than zero, then binds this namespace to the slice
+    ///   `[start..start + prefix_len]` in the buffer.
+    /// * else defines the current default namespace.
+    prefix_len: usize,
+    /// The length of a namespace name (the URI) of this namespace declaration.
+    /// Name started just after prefix and extend for `value_len` bytes.
+    ///
+    /// The XML standard [specifies] that an empty namespace value 'removes' a namespace declaration
+    /// for the extent of its scope. For prefix declarations that's not very interesting, but it is
+    /// vital for default namespace declarations. With `xmlns=""` you can revert back to the default
+    /// behaviour of leaving unqualified element names unqualified.
+    ///
+    /// [specifies]: https://www.w3.org/TR/xml-names11/#scoping
+    value_len: usize,
+    /// Level of nesting at which this namespace was declared. The declaring element is included,
+    /// i.e., a declaration on the document root has `level = 1`.
+    /// This is used to pop the namespace when the element gets closed.
+    level: i32,
+}
+
+impl NamespaceEntry {
+    /// Gets the namespace name (the URI) slice out of namespace buffer
+    ///
+    /// Returns `None` if namespace for this prefix was explicitly removed from
+    /// scope, using `xmlns[:prefix]=""`
+    #[inline]
+    fn namespace<'b>(&self, buffer: &'b [u8]) -> Option<&'b [u8]> {
+        if self.value_len == 0 {
+            None
+        } else {
+            let start = self.start + self.prefix_len;
+            Some(&buffer[start..start + self.value_len])
+        }
+    }
+
+    /// Check if the namespace matches the potentially qualified name
+    #[inline]
+    fn is_match(&self, buffer: &[u8], qname: &[u8]) -> bool {
+        if self.prefix_len == 0 {
+            !qname.contains(&b':')
+        } else {
+            qname.get(self.prefix_len).map_or(false, |n| *n == b':')
+                && qname.starts_with(&buffer[self.start..self.start + self.prefix_len])
+        }
+    }
+}
+
+/// A namespace management buffer.
+///
+/// Holds all internal logic to push/pop namespaces with their levels.
+#[derive(Debug, Default, Clone)]
+pub struct NamespaceResolver {
+    /// A stack of namespace bindings to prefixes that currently in scope
+    bindings: Vec<NamespaceEntry>,
+    /// The number of open tags at the moment. We need to keep track of this to know which namespace
+    /// declarations to remove when we encounter an `End` event.
+    nesting_level: i32,
+}
+
+impl NamespaceResolver {
+    /// Begins a new scope and add to it all [namespace bindings] that found in
+    /// the specified start element.
+    ///
+    /// [namespace binding]: https://www.w3.org/TR/xml-names11/#dt-NSDecl
+    pub fn push(&mut self, start: &BytesStart, buffer: &mut Vec<u8>) {
+        self.nesting_level += 1;
+        let level = self.nesting_level;
+        // adds new namespaces for attributes starting with 'xmlns:' and for the 'xmlns'
+        // (default namespace) attribute.
+        for a in start.attributes().with_checks(false) {
+            if let Ok(Attribute { key: k, value: v }) = a {
+                if k.starts_with(b"xmlns") {
+                    match k.get(5) {
+                        None => {
+                            let start = buffer.len();
+                            buffer.extend_from_slice(&*v);
+                            self.bindings.push(NamespaceEntry {
+                                start,
+                                prefix_len: 0,
+                                value_len: v.len(),
+                                level,
+                            });
+                        }
+                        Some(&b':') => {
+                            let start = buffer.len();
+                            buffer.extend_from_slice(&k[6..]);
+                            buffer.extend_from_slice(&*v);
+                            self.bindings.push(NamespaceEntry {
+                                start,
+                                prefix_len: k.len() - 6,
+                                value_len: v.len(),
+                                level,
+                            });
+                        }
+                        _ => break,
+                    }
+                }
+            } else {
+                break;
+            }
+        }
+    }
+
+    /// Ends a top-most scope by popping all [namespace binding], that was added by
+    /// last call to [`Self::push()`].
+    ///
+    /// [namespace binding]: https://www.w3.org/TR/xml-names11/#dt-NSDecl
+    pub fn pop(&mut self, buffer: &mut Vec<u8>) {
+        self.nesting_level -= 1;
+        let current_level = self.nesting_level;
+        // from the back (most deeply nested scope), look for the first scope that is still valid
+        match self.bindings.iter().rposition(|n| n.level <= current_level) {
+            // none of the namespaces are valid, remove all of them
+            None => {
+                buffer.clear();
+                self.bindings.clear();
+            }
+            // drop all namespaces past the last valid namespace
+            Some(last_valid_pos) => {
+                if let Some(len) = self.bindings.get(last_valid_pos + 1).map(|n| n.start) {
+                    buffer.truncate(len);
+                    self.bindings.truncate(last_valid_pos + 1);
+                }
+            }
+        }
+    }
+
+    /// Resolves a potentially qualified **element name** or **attribute name**
+    /// into (namespace name, local name).
+    ///
+    /// *Qualified* names have the form `prefix:local-name` where the `prefix` is
+    /// defined on any containing XML element via `xmlns:prefix="the:namespace:uri"`.
+    /// The namespace prefix can be defined on the same element as the element or
+    /// attribute in question.
+    ///
+    /// *Unqualified* attribute names do *not* inherit the current *default namespace*.
+    ///
+    /// # Lifetimes
+    ///
+    /// - `'n`: lifetime of an attribute or an element name
+    /// - `'ns`: lifetime of a namespaces buffer, where all found namespaces are stored
+    #[inline]
+    pub fn resolve<'n, 'ns>(
+        &self,
+        qname: &'n [u8],
+        buffer: &'ns [u8],
+        use_default: bool,
+    ) -> (Option<&'ns [u8]>, &'n [u8]) {
+        self.bindings
+            .iter()
+            .rfind(|n| n.is_match(buffer, qname))
+            .map_or((None, qname), |n| {
+                let len = n.prefix_len;
+                if len > 0 {
+                    (n.namespace(buffer), &qname[len + 1..])
+                } else if use_default {
+                    (n.namespace(buffer), qname)
+                } else {
+                    (None, qname)
+                }
+            })
+    }
+
+    /// Finds a [namespace name] for a given qualified **element name**, borrow
+    /// it from the specified buffer.
+    ///
+    /// Returns `None`, if:
+    /// - name is unqualified
+    /// - prefix not found in the current scope
+    /// - prefix was [unbound] using `xmlns:prefix=""`
+    ///
+    /// # Lifetimes
+    ///
+    /// - `'ns`: lifetime of a namespaces buffer, where all found namespaces are stored
+    ///
+    /// [namespace name]: https://www.w3.org/TR/xml-names11/#dt-NSName
+    /// [unbound]: https://www.w3.org/TR/xml-names11/#scoping
+    #[inline]
+    pub fn find<'ns>(&self, element_name: &[u8], buffer: &'ns [u8]) -> Option<&'ns [u8]> {
+        self.bindings
+            .iter()
+            .rfind(|n| n.is_match(buffer, element_name))
+            .and_then(|n| n.namespace(buffer))
+    }
+}
